+++ AlgorithmComparison/src/main/java/algorithmcomparison/algorithms/SortAlgorithm.java
@@ -4,6 +4,6 @@ public interface SortAlgorithm {
 
 	public void sort(int[] array);
 	public int getNumberOfKeyComparisons();
-	public long getRuntimeMilliseconds();
+	public int getRuntimeMilliseconds();
 	public void reset();
 }
